fix:  stores 테이블의 storeId 타입을 int에서 long으로 변경
- 공공데이터포털의 상품 데이터에서 entpId가 비정상적으로 높은 데이터 하나 관측
- 이에 따라 타입 변경

+++ backend/src/main/java/com/storerader/server/common/entity/StoreEntity.java
@@ -21,7 +21,7 @@ public class StoreEntity {
     private Long id;
 
     @Column(name = "store_id", nullable = false)
-    private Integer storeId;
+    private Long storeId;
 
     @Column(name = "store_name", nullable = false)
     private String storeName;
@@ -57,7 +57,7 @@ public class StoreEntity {
     private OffsetDateTime updatedAt;
 
     public StoreEntity(
-            Integer storeId,
+            Long storeId,
             String storeName,
             String telNo,
             String postNo,
